I hate this

+++ src/libpngchunkdec/Helpers.cs
@@ -4,14 +4,14 @@ namespace libpngchunkdec
 {
     public static class Helpers
     {
-        public static byte[] Read(this Stream stream, int length)
+        public static byte[] Read(this Stream stream, long length)
         {
             byte[] buff = new byte[length];
             stream.Read(buff, 0, buff.Length);
             return buff;
         }
 
-        public static byte[] ReadMotorola(this Stream stream, int length)
+        public static byte[] ReadMotorola(this Stream stream, long length)
         {
             byte[] buff = new byte[length];
             stream.Read(buff, 0, buff.Length);
